listing restaurants

+++ app/src/main/java/com/renanrhoden/wheretolunch/repositories/PlacesRepository.kt
@@ -1,7 +1,7 @@
 package com.renanrhoden.wheretolunch.repositories
 
+import com.renanrhoden.wheretolunch.model.Result
 import com.renanrhoden.wheretolunch.service.PlacesApi
-import com.renanrhoden.wheretolunch.service.Result
 import io.reactivex.Single
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
@@ -9,7 +9,7 @@ import retrofit2.Retrofit
 
 class PlacesRepository(private val retrofit: Retrofit) {
 
-    fun getRestaurants(latitude: Long, longitude: Long): Single<Result> {
+    fun getRestaurants(latitude: Double, longitude: Double): Single<Result> {
         return retrofit.create(PlacesApi::class.java)
             .getPlaces("$latitude,$longitude")
             .subscribeOn(Schedulers.newThread())
